requirements.txt
pandas
numpy
requests
pyarrow
matplotlib
seaborn
jupyterlab
tenacity
networkx
python-dotenv
notebook
ipykernel